+++ internal/translation/models.go
@@ -3,7 +3,7 @@ package translation
 import "time"
 
 type Translation struct {
-	ID          uint64 `gorm:"primaryKey"`
+	ID          int `gorm:"primaryKey"`
 	Text        string
 	Translation string
 	LangFrom    string
@@ -14,8 +14,8 @@ type Translation struct {
 }
 
 type Score struct {
-	UserID        int    `gorm:"primaryKey"`
-	TranslationID uint64 `gorm:"primaryKey"`
+	UserID        int `gorm:"primaryKey"`
+	TranslationID int `gorm:"primaryKey"`
 	Score         int
 	CreatedAt     time.Time
 	UpdatedAt     time.Time
